requirements.txt
openai>=1.12.0
stravalib>=0.10.5
python-dotenv>=1.0.0
requests>=2.31.0
